requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pymongo==4.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.1
